Summary: Adding new CAENAcquisitionTemplate for fast acquisition code prototyping
Details:

CAENAcquisitionTemplate is intended to be a fully documented,
copy-and-develop acquisition code for use with the CAEN V1718 and
V1720 boards. The purpose is to educate the user on how such
acquisition is accomplished as well as to provide a complete and
mature foundation for developing their own codes or fast prototype
codes for testing. CAENAcquisitionTemplate includes a mature build
system that handles dependencies and compilation as well as the basic,
modular layout for acquisition.