Finished prototype of ADAQReadoutManager class
Details:

The ADAQReadoutManager prototype has been completed. Without having
done any testing/debugging, it appears to provide all the basic
functionality need for complete readout of CAEN hardware using the
ADAQAcquisition code, which is its primary purpose. Testing soon to
commence, and there are likely to be some fundamental changes to
accomodate inevitable issues. In particular, the decision regarding
whether to contain the Waveforms and WaveformData as class members of
the ADAQReadoutManager class or to merely set pointers/address. of those
from within the ADAQAcquisition AAAcquisitionManager class