Updates for information readout
Details:

A lot of work to the ADAQReadoutInformation class, mainly in new class
data members and in set/get methods for all data

Work has also been done to add information functionality to
ADAQDigitizer such that the important board information may be
retreived with set get methods. Some new structure have been added to
support the automatic setting of various important data members based
on the type of digitization device that is opened.